refactor: separate view model's responsibility
